Removed hphp/util/db-mysql.(cpp|h)
This was leftover from an ancient abstraction requirement that is no longer present.
